# backend/docs/swagger.yaml
definitions:
  handlers.CreateRoomInput:
    properties:
      capacity:
        type: integer
      description:
        type: string
      name:
        type: string
    required:
    - capacity
    - name
    type: object
...
  handlers.UpdateRoomInput:
    properties:
      capacity:
        type: integer
      description:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/rooms:
    get:
      consumes:
      - application/json
      description: Get list of rooms with optional pagination and name filter
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Room name filter
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all rooms
      tags:
      - room
    post:
      consumes:
      - application/json
      description: Create a new room
      parameters:
      - description: Room info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRoomInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create room
      tags:
      - room
  /api/rooms/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a room by ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete room
      tags:
      - room
    get:
      consumes:
      - application/json
      description: Get detail of a room by ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get room detail
      tags:
      - room
    put:
      consumes:
      - application/json
      description: Update an existing room by ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      - description: Room info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateRoomInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update room
      tags:
      - room
swagger: "2.0"